+++ core/internal/cltest/factories.go
@@ -645,8 +645,8 @@ func MustInsertOffchainreportingOracleSpec(t *testing.T, store *strpkg.Store, de
 		P2PPeerID:                              models.PeerID(DefaultP2PPeerID),
 		P2PBootstrapPeers:                      []string{},
 		IsBootstrapPeer:                        false,
-		EncryptedOCRKeyBundleID:                DefaultOCRKeyBundleIDSha256,
-		TransmitterAddress:                     DefaultKeyAddressEIP55,
+		EncryptedOCRKeyBundleID:                &DefaultOCRKeyBundleIDSha256,
+		TransmitterAddress:                     &DefaultKeyAddressEIP55,
 		ObservationTimeout:                     0,
 		BlockchainTimeout:                      0,
 		ContractConfigTrackerSubscribeInterval: 0,
